requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
html5lib==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.5
psycopg2-binary==2.8.5
regex==2020.4.4
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tqdm==4.45.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1